requirements.txt
celery==5.2.7
Django==4.1.2
djangorestframework==3.12.4
django-cors-headers==3.8.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-decouple==3.6
redis==4.3.4
whitenoise==6.6.0